Refactor scope classes into separate modules
